Many changes: Log, index type, append entries proto etc
- Make a Log struct which will be responsible for keeping track of the
  log. It will write the end marker as well as keep track of any
  positions of the state file
- Type of index will be used as usize internally, whereas it'll be used
  as u64 when talking amongst each other
- append entries proto will have entries that contain message as well as
  term, this helps a leader in committing entries from previous terms as
  well

+++ src/consensus/mod.rs
@@ -74,7 +74,7 @@ pub struct Node<SFile: StateFile> {
 
 pub struct NodeServer<SFile: StateFile> {
     node: Arc<Node<SFile>>,
-    last_log_index_informer: watch::Sender<u64>,
+    last_log_index_informer: watch::Sender<usize>,
 }
 
 /// Represents information about a peer node that a particular node has and owns
